03-agents/pyproject.toml
[project]
name = "agents-module"
version = "0.1.0"
description = "Modulo 3: Agentes y LLMOps aplicado con OpenAI GPT-5 mini."
requires-python = ">=3.12"
dependencies = [
  # Existentes
  "openai>=1.60.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.8.2",
  "typer>=0.12.5",
  "rich>=13.7.1",
  # Nuevas: LangChain + LangGraph
  "langchain>=0.3.0",
  "langchain-core>=0.3.0",
  "langchain-openai>=0.3.0",
  "langchain-community>=0.3.0",
  "langchain-text-splitters>=0.3.0",
  "langgraph>=0.4.0",
  # Vectores y embeddings
  "chromadb>=0.5.0",
  "tiktoken>=0.7.0",
  # Analisis y visualizacion
  "matplotlib>=3.8.0",
  "pandas>=2.1.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.3.0",
  # Resiliencia
  "tenacity>=8.2.0",
  # Notebooks
  "ipykernel>=6.29.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=4.1.0",
  "ruff>=0.8.0",
]

[project.scripts]
llmops-run = "llmops.main:app"

[tool.uv]
package = true

[tool.setuptools]
package-dir = {"" = "LLMops/src"}

[tool.setuptools.packages.find]
where = ["LLMops/src"]

[tool.pytest.ini_options]
testpaths = ["LLMops/tests", "tests"]
python_files = ["test_*.py"]
addopts = [
  "-v",
  "--strict-markers",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "W"]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"
